+++ src/components/QuoteSettings/Quotes.js
@@ -3,7 +3,7 @@ import React from 'react';
 import Quote from './Quote';
 import './QuoteSettings.scss';
 
-const Quotes = ({ quotes }) => {
+const Quotes = ({  quotes }) => {
   const quoteItems = quotes.map((quote, i) => {
     return (<Quote key={i} quote={quote} />);
   });
